% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjHR.R
\name{adjustment}
\alias{adjustment}
\title{Cox PH – Taylor expansion adjustment for Long-term survival data}
\usage{
adjustment(
  HR_cox,
  HR_cox_CI,
  s1mix.chosen,
  s0mix.chosen,
  pi1.est,
  pi0.est,
  m = NULL,
  cure.show = TRUE
)
}
\arguments{
\item{HR_cox}{Input. Hazard ratio obtained from Cox PH model.}

\item{HR_cox_CI}{Input. Confidence interval of the hazard ratio obtained from Cox PH model.}

\item{s1mix.chosen}{Input. Survival probabilities of Arm 1 estimated by KM at the chosen time.}

\item{s0mix.chosen}{Input. Survival probabilities of Arm 0 estimated by KM at the chosen time.}

\item{pi1.est}{Input. Proportion of short-term survivors (uncured proportion) in Arm 1.}

\item{pi0.est}{Input. Proportion of long-term survivors (uncured proportion) in Arm 0.}

\item{m}{Input (Optional). The polynomial order used in Taylor approximation.
\emph{m} is selected automatically if m = NULL (default).}

\item{cure.show}{Input (Optional). Indicate whether or not to show the estimated proportions of long-term survivors
The default is TRUE.}
}
\value{
\item{Adj.before.after}{Results before and after Cox-TEL adjustment: Cox HR, Cox-TEL HR, and Cox-TEL DP.}

\item{Proportion.LTS}{Estimated proportions of long-term survivors with confidence intervals.}
}
\description{
An adjustment for Cox hazard ratio and an estimation for difference in proportions.
}
\examples{
library(CoxTEL)
s1mix.chosen <- c(0.48,0.39,0.35,0.35)
s0mix.chosen <- c(0.36,0.28,0.26,0.25)
pi1.est <- 0.65; pi0.est <- 0.75 # Proportion of short-term survivors
HR_cox <- 0.71; HR_cox_CI <- c(0.51, 0.91)
adjustment(HR_cox, HR_cox_CI, s1mix.chosen, s0mix.chosen, pi1.est, pi0.est)

}
\references{
C.Y. Hsu, E.P.Y. Lin and Y Shyr. (2021). Development and Evaluation of a Method to
Correct Misinterpretation of Clinical Trial Results With Long-term Survival.
JAMA Oncol. doi:10.1001/jamaoncol.2021.0289.
}
