% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCP2cure.R
\name{plotCP2cure}
\alias{plotCP2cure}
\title{Plot conditional probability of Pts belonging to long-term survivors}
\usage{
plotCP2cure(CT.object, timeunit = 6, alpha.f = 0.3, ...)
}
\arguments{
\item{CT.object}{Input. Object of \emph{adjustment}.}

\item{timeunit}{Input. Time scale.}

\item{alpha.f}{Input. Alpha transparency in \emph{graphics::polygon}.}

\item{...}{Input. Arguments in \emph{plot}.}
}
\description{
A figure function to show conditional probability of Pts belonging to long-term survivors.
}
\details{
P(Pts belongs to long-term survivors | T>t) = (1 - pi)/smix(t).
The confidence interval are defined as ((1 - pi.L)/smix(t), (1 - pi.U)/smix(t)).
}
\examples{
library(CoxTEL)
s1mix.chosen <- c(93, 83, 74, 66, 61, 56, 52, 49, 44, 42)/100
s0mix.chosen <- c(86, 75, 62, 55, 48, 42, 40, 36, 33, 33)/100
pi1.est <- 1 - 0.42; pi0.est <- 1 - 0.33 # Proportion of short-term survivors
HR_cox <- 0.68; HR_cox_CI <- c(0.53, 0.87)
CT.object <- adjustment(HR_cox, HR_cox_CI, s1mix.chosen, s0mix.chosen, pi1.est, pi0.est)
plotCP2cure(CT.object, xlab="Months")

}
