% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjHR.R
\name{adjustment}
\alias{adjustment}
\title{An adjustment for Cox hazard ratio and an estimation for difference in proportions.}
\usage{
adjustment(HR_cox, HR_cox_CI, s1mix.chosen, s0mix.chosen, pi1.est, pi0.est,
  m = NULL)
}
\arguments{
\item{HR_cox}{Input. Hazard ratio obtained from Cox PH model.}

\item{HR_cox_CI}{Input. Confidence interval of the hazard ratio obtained from Cox PH model.}

\item{s1mix.chosen}{Input. Survival probabilities of Arm 1 estimated by KM at the chosen time.}

\item{s0mix.chosen}{Input. Survival probabilities of Arm 0 estimated by KM at the chosen time.}

\item{pi1.est}{Input. Proportion of poor-responders (uncured proportion) in Arm 1.}

\item{pi0.est}{Input. Proportion of poor-responders (uncured proportion) in Arm 0.}

\item{m}{Input (Optional). The polynomial order used in Taylor approximation.
The default value is NULL that means \emph{m} is selected automatically.}
}
\value{
\item{HR_cox}{Cox hazard ratio you inputed.}

\item{HR_cox_CI}{Confidence interval of the Cox hazard ratio you inputed.}

\item{HR_adj}{Hazard ratio after adjustment.}

\item{HR_adj_CI}{Confidence interval of the after adjustment.}

\item{DP_adj}{Difference in proportions of the true responders.}

\item{DP_adj_CI}{Confidence interval of the difference in proportions of the true responders.}
}
\description{
A novel adjustment method for Cox proportional hazards model in data with long-term survival.
}
\examples{
library(AdjCOXPH)
s1mix.chosen <- c(0.48,0.39,0.35,0.35)
s0mix.chosen <- c(0.36,0.28,0.26,0.25)
pi1.est <- 0.65; pi0.est <- 0.75
HR_cox <- 0.71; HR_cox_CI <- c(0.51, 0.91)
adjustment(HR_cox, HR_cox_CI, s1mix.chosen, s0mix.chosen, pi1.est, pi0.est)

}
